requirements.txt
annotated-types==0.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
idna==3.10
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
